backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-jwt-extended==4.5.3
psycopg2-binary==2.9.11
google-genai==0.3.0
python-dotenv==1.0.0
gunicorn==21.2.0
bcrypt==4.1.2
requests==2.31.0
flask-bcrypt==1.0.1
